% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOTeR.R
\docType{data}
\name{df}
\alias{df}
\title{Default dataset.}
\format{
A default dataset includes:
\describe{
\item{.id}{device serial number}
\item{Index_level}{INdex of manipulated data}
\item{Moisture}{moisture data in percentage}
\item{Radius}{radius data in micrometers}
\item{T1}{Temperature 1 of TMS sensors}
\item{T2}{Temperature 2 of TMS sensors}
\item{T3}{Temperature 3 of TMS sensors}
\item{date_time}{Date and time stamp in UTC}
}
}
\usage{
df
}
\description{
Default dataset.
}
\keyword{datasets}
