% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOTeR.R
\name{PLOTeR}
\alias{PLOTeR}
\title{PLOTeR}
\usage{
PLOTeR()
}
\description{
\code{PLOTeR()} Web application for time series data visualization and processing.
}
\details{
Developed for processing of data from dendrometers and soil moisture sensors.Mostly adapted for data structure of Tomst sensors. Reads data.frame named "df" as default.
}
\examples{
PLOTeR()

}
