% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tomst.R
\name{read_tomst}
\alias{read_tomst}
\title{read_tomst}
\usage{
read_tomst(
  path,
  file,
  radius_units = c("auto", "um", "tomst"),
  interval = c("auto", "1 min", "5 min", "15 min", "1 hour"),
  TMS_calibration = c("Loamy_Sand_A", "Loamy_Sand_B", "Sandy_Loam_A", "Sandy_Loam_B",
    "Loam", "Sil_Loam", "Peat", "Sand"),
  delim
)
}
\arguments{
\item{path}{a character vector of full path name; the default corresponds to the working directory, getwd().}

\item{file}{a character vector of file names; e.g., c("data_92241809_2025_04_04_0.csv", "data_92241809_2025_04_04_1.csv"). If missing, all csv files in the working directory matching the name pattern "data_8digitserialnumber_*.csv" will be loaded.}

\item{radius_units}{the units of dendrometer radius used when downloading data using Tomst software. Use "auto" to automatically convert radius values into micrometers (default). Use either "tomst" or "um" to convert manually. Specifically, if the data were downloaded from Tomst software in micrometers, use "um", if data were downloaded in Tomst units, use "tomst". This function always attempts to convert data into micrometers.If values lower than 1270 are found together with "tomst" units, warning message will be displayed.The conversion formula from "tomst" units into micrometers is: micrometers = (tomst-1278)*(8890/(34000-1278))}

\item{interval}{parameter used to fill in missing data_stamps as the desired interval. "auto" detects the interval automatically. If multiple intervals are automatically detected, the prevailing one is used. Variables will contain NA for the filled in data_stamps. Shorter intervals than the desired one are kept in data.}

\item{TMS_calibration}{recalculation of Tomst TMS soil moisture count into volumetric soil moisture based on calibration equations (version available on TOMST.com 06.10.2025). Loamy_Sand_A is used as default.}
}
\value{
Data.frame with long type of data.
}
\description{
\code{read_tomst()} Returns data frame from csv files of Tomst structure.
}
\details{
Data are automatically named and converted into default structure.
}
\examples{
df <- read_tomst(radius_units = "tomst", delim = ".")
df <- read_tomst("data_94000000_0.csv", delim = ".")

}
